pyproject.toml
[tool.poetry]
name = "pynamodb-templates"
version = "0.1.0"
description = "PynamoDB template with common custom attributes"
authors = ["SeongHun Kim <shkim@deephearing.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pynamodb_templates"}]

[tool.poetry.dependencies]
python = ">=3.8"
pynamodb = "^5.3.4"
python-ulid = "^1.1.0"
pynamodb-attributes = "^0.4.0"
typing-extensions = "^4.4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=6"
docker = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
